register part complete

+++ domain/src/main/java/com/example/domain/usecase/CartSummaryUseCase.kt
@@ -6,5 +6,5 @@ import com.example.domain.repository.CartRepository
 class CartSummaryUseCase(
     private val cartRepository: CartRepository
 ) {
-    suspend fun execute(userId:Int) = cartRepository.getCartSummary(userId)
+    suspend fun execute(userId:Long) = cartRepository.getCartSummary(userId)
 }
