+++ domain/src/main/java/com/example/domain/usecase/DeleteProductUseCase.kt
@@ -6,5 +6,5 @@ import com.example.domain.repository.CartRepository
 class DeleteProductUseCase(
     private val cartRepository: CartRepository
 ) {
-    suspend fun execute(cartItemId: Int, userId:Int) = cartRepository.deleteItem(cartItemId,userId)
+    suspend fun execute(cartItemId: Int, userId:Long) = cartRepository.deleteItem(cartItemId,userId)
 }
